StepFunctions/BatchImageProcessing/lambdas/ProcessExifFunction/requirements.txt
Pillow==8.1.2 